+++ Map.cpp
@@ -12,13 +12,13 @@ initializeIslands(10, checkDuplicateCoordination(island_));
 std::size_t Map::randomGenerate(std::size_t min, std::size_t max){
     	std::random_device rd;
 		std::mt19937 mt(rd());
-		std::uniform_int_distribution<int>dist(min, max);
+		std::uniform_int_distribution<size_t>dist(min, max);
 		std::size_t random_pos = dist(mt);
         return random_pos;
 }
 
 void Map::initializeIslands(const std::size_t numberOfIslands, bool ifDuplicate){
-    for (int i = 0; i < numberOfIslands; i++){
+    for (size_t i = 0; i < numberOfIslands; i++){
         if (ifDuplicate == true){
             i--;
         } else {
